Fixed issue with delayed API post
Reworked the item API and axios post, added event.preventDefault() as it was having issues without it. Added auth check on landing page to redirect to the dashboard(HUB) when logged in. Also added auth to the dashboard to send the user to the landing page if not logged in.