Added Celestial Tower component, CSS tweaks, file name changes
Added the Celestial Tower component and route to be a Dungeon of sorts where I can build and test out the combat system. Did some minor CSS tweaks on the HUB to fit it all on a 1920 x 1080 monitor when full screen. Imported the google fonts I will be using, but havent implemented it yet. Renamed Dashboard to HUB and renamed all routes for it.