Added the Combat component and reposition mechanics
I created the Combat component to hold all combat. I'm currently using static data in the client for abilities that I'll eventually move to the server, allowing users to unlock abilities and save a template of abilities to use within the component. Reposition elements are working properly and update the page when the player moves positions.