Reworked the way location is saved, removed calls for getUser and added a new saveUser api post
I was having issues previously with lengthy async calls. I decided to no longer retrieve the user data from the database before calls and only happens once on login.

Location is now saved to the user's jwt token, and the application saves the user's character data to the database after modifying it. I did it this way so that I wasn't constantly saving a single value (location) to the database. It'll clear up some issues later down the road and is a good starting point to refactoring the current system.