Re-added redirect functionality to the HUB and Celestial Tower
With the rework of the new save location function, I've now got the redirects working as intended. If the player is in the HUB ("/HUB"), and tries to go to the Celestial Tower ("/HUB/CelestialTower") by typing it in the URL, the player will be redirected back to the HUB as it is the current location saved. This prevents players from trying to go to different locations by knowing the URL, but without actually moving their character there.