Refactored setLocation into setLocalUser
I was having issues running multiple different token modifications using multiple api routes. I repurposed the setLocation api that originally was just being used to set the location into a saveLocalUser api that will save the user data that is given to it.

This allows me to use the same api route to save any modification to the character to the local data and another route that saves the current local data to the database. I believe this is a very streamlined approach to the issue, and will make further modifications much easier to perform.