Modified saveuser api
The save user api was only pulling the character's location data, modified it to include all character data now that it's being used for the whole character.

At this point, saveUser() is working. It pulls the current local data and saves it to mongoDB. saveLocalUser() is used when I need to modify the local user's data, such as changing their location or adding an item. This may also be used later down the road when I need to implement abilities.