Redesigned the combat system
I redesigned the combat system to be reusable and called from any location I desire.

If I want to initiate combat, I redirect the player to the CombatController component. This component does a lot of things. It prepares the user's data (Abilities, Preferred Position, etc...) and the monster data to be passed into the specific combat component that is chosen via the play's current location.

Example: If the player is in the Celestial Tower, the combat controller will redirect the player to CelestialTower/Combat, which loads the CombatCelestialTower component. This component will be a unique combat component, one made for each zone that will have combat.

The combat component will receive the props from the combat controller, and then play out the combat.... which needs to be completed at a later date.

All of this redesign is to allow me to spawn a new combat unique to a zone, and allowing me to spend less time on each individual combat component, as all it needs is the images. I could even go one step further with this, and refactor the current CombatCelestialTower component into a generic CombatComponent by passing in more information for the component such as the side images.

I will most likely do this in a second pass on the combat overhaul prior to working on the combat mechanics itself. I'm hoping to have one combat component that is called and generically used by all zones, with the controller passing the unique data for the component. This will make it much easier to write the combat logic, as it'll be contained within a single component.

I reorganized the combat file structure as I started using external file functions as they are being used in multiple components. CombatFunctions holds all the external functions and CombatComponents will most likely hold the single combat component.