Fixed dispatch issue and react warning in Navbar
The dispatch issue I was having was due to redux-thunk not pulling in dispatch for one reason or another. I decided to instead have the async function return the action itself, so that I can map dispatch to props and run the dispatch within the component itself. I'm not sure if this is the best use case in this scenario, but from what I can tell, everything is working properly and no data is being lost anymore.

I fixed a warning in the Navbar where react was getting fussy about my if/else statements within render. I moved them to the navbar's state constructor. I'll work on the navbar more when it comes time to implement the combat logic.