Condensed Nav Menus
I've removed the two buttons for Character Stats and Chat from the Navbar. In their place, I've added a menu button that pulls out a curtain menu with Character and Chat on it. These new links open the previous components for Character and Chat.

Added react-draggable npm package to add click/drag functionality to the panels, so that the player may move them around as they wish.