Added Forgot Password system
Users now can click on "Forgot My Password" on the login page to be sent to another page where they can input their email. Once the email is inputted, they can click send email to recieve an email from the application's gmail account that contains a link with a token.

Upon clicking that link, the application will check the link's token vs the db's token and if the token is correct, then it'll render the input box to reset the user's password. Upon submission, the user's password is hashed/encrypted and stored and ready to be used.

In this process, I've also added an updatePassword route that I can use for a password reset page while logged into the app.