Created Abilities.js
Abilities.js will be the new file location for all player abilities. This includes damage, chainer, finisher, reposition, heal, generic and eventually magic abilities. I will build an api route that the retrieves the abilities in JSON format to allow the users to view the ability data in a list format.