Added API Route, Added AbilityPanel
I added the API Route to retrieve the abilty JSON from the server, this will let me show all the abilities on the AbilityPanel. The AbilityPanel will also be where the user will see the experience they have collected for the weapon types, and the abilities for the weapons that they can unlock.