Ability Panel Integration
I fleshed out the Ability Panel to now take in the abilities from the server's api. It then allows the user to click through the buttons on the left to view the abilities tied to that weapon(or utility). I will add functionality for users to see what abilities they already have unlocked and the ability to unlock abilities in the future.