Ability Tooltips
The ability tooltips have been fleshed out, showing all the data for the corresponding skill that was clicked on. The user can also learn the skill or close the tooltip window(or pick another skill). Next is to add functionality for the user to know what skills they have learned, and remove the ability to learn a skill if they already own it.