Experience and Ability Unlocks added.
The user can now view the amount of exp they have for each weapon type in the ability panel. They can click the ability they want to learn to view the tooltip. They then can click "Learn" to learn the ability which removes the xp from the character required to learn it.

The ability panel then shows which abilities the user has unlocked by showing " // Unlocked" after the ability name. The learn button on the ability tooltip is also removed if already learned.

The action that performs the saving of the ability's id to the character's unlocked abilities attribute checks to see if the user already learned it, and if so, it just skips it. This is added protection just incase the button fails for whatever reason.