Combat Prefs started
I've added the combat prefs panel where the user will choose their abilities to go into each slot (Positions 1-6 basic attacks, chainers, finishers, and their utility abilities such as repositions and heals).

I currently have weapon selection using similar functions to the ability panel to show the weapon abilities the user has learned.

I'm not sure how i'm going to go about the user selecting the skill, but i'm thinking i'll just make a popup for each ability slot that lets the user choose the ability to go into it. That way, I can have a single component take in the skill position, and will show the abilities that go into that slot. It can then save the ability to the slot in the user's db so they can use it for combat.