Finished Combat Prefs Panel But Encountered Major Issue
I finished implementing the combat prefs panel, and it's working as intended with a weapon swap in mind(Allowing users to swap between two weapons and save the skill data for each one).

This all works fine, the issue is this. I've been saving character data in cookies so that the data can be persistant for the user. The issue with this is that i've hit max cookie data limit apparently, so I'm going to need to rework how data is saved to redux.