Fixed issues with the change to saving local user
I had to do major sweeping changes across the whole application. The User Schema had some issues itself and needed to be fixed in some areas, mainly pertaining to default values of properties.

I had to fix a couple things in most components, such as holding a default state if the user is not loaded, as the user is now being loaded upon loading Wrapper.js.

I'll look more into this later before continuing with developing other things. I want to have this fleshed out and working flawlessly.

One issue at the moment is the application requiring a refresh of the page upon load to get the user's data.