Fixed Issues
I was having an issue with a couple different things after migrating the character data away from the cookies.

First issue: I noticed that when I would grab the user data, it looked fine, but after modifying it, it would miss empty objects. Fixed this by modifying the user shema's default values and adding an option at the end.

Second Issue: I was not successfully retrieving and setting the user data every time the user loaded up the page.
Fixed this by using both componentDidUpdate and componentDidMount to grab the user data both on mount and update if the user does not currently have their character information.

I've looked into hooks and i'm tempted to rewrite the structure of the current site  to use hooks and functional components instead of class based, but that would require me to set up props due to them not being connected. Will look into it.

Added lodash to clone the redux object.
Cleaned up console logs.