Added Comments in prep for combat update
I plan on reworking the Combat Controller soon, I have almost everything in place for it to work besides user's preferred location in the combat prefs. Once I have that finished, I should be able to flesh out the majority of the combat controller so it is ready to work with any location data.