User can now change their preferred position
The user can now click on the "Preferred Position" button within the combat prefs, this opens a new popup panel with 1-6 as the position choices. This saves that position as their preferred position that they will start out at in combat.

This was the last feature needed to start working on the combat controller.