Fixed Combat Pref Bug
Clicking on reposition forward or reposition backward would result in the user seeing both forward and backward reposition skills in the popup. Added another check in the filter so that reposition forward will only see "Charge" as the current forward repositions skill, and "Dodge Back" as the backward skill.

I removed the "ready" key from the combatPrefs object, mainly because there will be no need to check if the player's combat pref skills dont contain skills. The original intention for this value was to swap to true when the player fills out all combat prefs with skills, but  I plan on having a new player enter the game, choose a weapon, and then have the combat prefs fill out with a basic set of skills. This makes it so the player will ALWAYS have skills inside combatPrefs.