Fixed loading the character data
The Wrapper Component will load after either the user logs in, or if the user returns after closing the browser with an active token.

It then takes the user's token data, and looks for their character in the Mongo Database. It will then return that data, dispatch it to Redux to give the character data, and then upon component update, it will check the character data and make sure it has data in it, and will then set the state for characterValid and characterUploaded to true.

When characterValid is false, the Wrapper shows a "Loading Character" element, and if it is true, it then allows the user into the HUB, which then will redirect the user to their saved location.

-------------

The HUB and Celestial Tower now have their location redirects uncommented. They are currently working very well. This now ensures that I can keep the user in the location they are saved on the database, removing the need to worry about user's cheating through URL manipulation.