Combat Prefs Component Update
Combat Prefs previouly required the refresh button to show changes to the user's combat prefs as they updated it. I didn't really like that, since the user should automatically view the change upon selection.

I wrote a small series of checks that has three stages: "Initial", "Open", and "Closed". The "Initial" state is for when the user does not have an ability/position panel open and the combat prefs is the only open panel. When the user chooses an ability or position to change, the flag gets set to "Open", and when the user selects their desired skill/position, the flag gets changed to "Closed". This makes it so I can update the page upon "Close" with the new combat prefs.