Backend Item Generation Post Working
I've worked on the item generation a bit and I can now generate an item from the server via a post API from the client. It's very basic right now and only generates tier one dagger weapons, but all I need to do now is to work on the meat of the item generation.

I'm planning on getting the basic items generated first (Tier one weapons and tokens), so I can give users items from the combat that i'll be working on later.