Added Skeleton Token to item generation
Rearranged the propertys added for item generation. It's now arranged in such a way that there are 6 different properties an item will have. Name/Type/Sub Type/Image/Tradeable/ID. These properties apply to all items, and thus are ran first in the generation.

The tokens will then have a count applied so that they are stackable in the inventory, and if the user generates another token after already having one, I'll just apply a +1 to the token they already have in their inventory instead of using the newly generated token.

If the generated item is a weapon, the item generation read's the blueprint's min and max damage values and then randomizes the flat damage amount the weapon will have. After the damage, it then calculates the rarity of the item, which contains a 1-10000 roll for each rarity upgrade, needing a more difficult check with each roll. Rarity does not do anything at this moment, but this allows me to start to get a feel of drop rates as I test the item generation.