Updated Abilities to have multiplier values
I reworked the abilities and removed the min/max values and replaced them with percentage multipliers for easier balancing in the future. Ran into a weird issue with auth and after some tinkering I fixed it.

While fixing this, I ran into a new issue. Whenever the user clicks logout, The application cannot find the user's ID, obviously due to logging out and the auth token no longer being in effect. This causes the wrapper to default to the Loading Character page as it has no valid character. I will need to figure out a way to path the user to the login page without causing this strange behavior.