Fixed Login/Logout issue
I was having an issue with some of my components rendering regardless if the user was logged in or logged out. This was due to me having the Wrapper component output a "Loading Character" placeholder text for a loading screen when the app was pulling character data.

After some tinkering, I was able to piece together a temporary solution. It's a bit messy at the moment, and i'm sure i'll return to it later to clean it up.

It's working for now, which will allow me to continue working on some basic item and equipment stuff. After that, I can work on the actual combat system once I have the ability for the user to equip and unequip weapons.