Item Fragmentation Added
The user may now click the "Fragment" button on an item in the inventory. This button destroys the item and gives the user Bound Fragments based off the item's fragment data. This will be the main means for the user to get rid of items in their inventory and will give them currency in return.