Added equipping a weapon
The inventory panel now supports equipping a weapon into the weapon one slot. This is the minimal amount of equipment swapping I need at the moment to get combat functioning. The user can now equip a weapon from the inventory and swaps it with the currently equipped weapon.

This will be expanded at some point.

During this, I found an error within combat prefs that I will need to fix.