Cleaned up Combat, Fixed issue with weapons
I started cleaning up the combat code to remove some out of date code I was no longer using, in preparation for working fully on creating the combat.

In addition to this, when creating the equipping system, I mistyped and created an issue when saving weapons. The equipping now properly saves the weapon to the correct position in the character data. This was also fixed in combat prefs.