Combat Positioning and Ability Interactions
The combat now properly pulls in all abilities the user uses. Basic abilities enable chainers, chainers enable finishers. This is all reflected in the user interface and abilities are properly showing in the correct positions and working as intended. This set was vital, as now I can use the damage values on abilities when clicked on.