Cleanup, UI Responsiveness, Combat Update
I've gone ahead and looked at the majority of the files. Cleanup consisted of:

Removing console.logs or commenting out ones I'll use in the future.
Removed unneeded code and old commented code.
Added some documentation comments discussing functions, how they work, and what they are used for.

UI Responsiveness:
For a long time, i've never really worked on the navbar and the occompanying UI elements such as the health bar and inventory. I've added some new state variables that I'm using to trigger updates on both the Wrapper and Navbar.

The Wrapper now can give any child component the updatePreviousAction() function. This function only changes the state for wrapper, by giving it the current action ("Heal", "Gained Item", "Won Combat"...etc), and attaches a unique ID to the end of the string, so that it is always unique. This allows me to "force" update Wrapper, and any other component to trigger a refresh allowing me to give the user updated visuals.

I am using this currently for:
Navbar Health Bar
Inventory Panel (Gaining an item and fragmenting an item).
Heal button in HUB

Combat:

I left combat at a pretty good point last time, having set up the user's abilities and having the proper position movement added. I went ahead and implemented damage to the user and the enemy. When the user defeats an enemy, the user wins the combat and is sent back to the previous location (Currently the Celestial Tower). When the user loses health, the navbar is updated to reflect this.

Things to be added to combat:
Item Rewards
Experience Rewards
Fragment Rewards
Enemy Movement and Positioning
Enemy Special Attacks
Implementing Stats(Crit, Initiative?, Others.)

Misc Additions:

Lodash (for clonedeep)