ZoneController, setSubLocation, and WorldZones
This commit consists of the first attempt at creating a ZoneController and setting up the infrastructure to add more zones in the future.

Currently, the ZoneController will take in two user properties. Location and SubLocation. Location will be the central location the user is in. It will consist of unique zones: HUB, Celestial Tower, and Major zone data such as Crystal Forest.

The location attribute lets me figure out where exactly the user is located to load the correct component.

If the user clicks on the Crystal Forest button inside the HUB, the user's location (CrystalForest) and subLocation (Spire Path, the entry zone into Crystal Forest) will be saved and they will be redirected to the ZoneController.

The ZoneController takes in both these variables, and loads up the correct component (CrystalForest). In addition, the ZoneController reads the subLocation (Spire Path), and retrieves the zone's data which consists of the Exits that zone has, the Name of the zone, and will include levels of monsters, possible unique loot and other data relating to that individual zone.

This setup allows me to build one component per major location such as Crystal Forest that holds many different smaller zones that the user may then navigate through.

Example: HUB -> Crystal Forest(Spire Path) -> Crystal Forest(Ruined Stairway) -> Crystal Forest(Shimmering Rise) -> Crystal Forest(Petrified Gate).

With this setup, I'll be able to modify what is included in each subLocation by adding new properties to the WorldZoneData file.

At this moment, I'm not seeing any major glaring issues. There might be some that arise regarding reading and setting the locations properly when leaving or entering major locations such as the HUB.

Implementation of combat should be easy, as the CombatController only needs to be updated to include these locations and sublocations to load up the correct monsters for that zone.