Zones Added, Fixed Potential Future Issue
All of the remaining Crystal Forest zones have been added. Zone transitions have been tested and are working fine.

I've decided to remove the function that saves the local user to the database when transitioning between zones. This potentially could lead to exploits. The reasoning behind this is the fact that with every save going into a zone, it'll clog the server with many attempts to save to the database, causing unneeded lag requiring a refresh.

This will be fixed by saving after a user does something in a zone such as combat, or collecting resources.

At this point in time, the ZoneController seems to be functioning as expected. I'm able to add new zones quickly by adding an object for the new zone and exits to/from other zones. I will also be able to add unique data to each zone as needed. Such as requesting a unique boss monster, or unique drops to that specific zone.

I'm also at the point where once combat is added, I can start work on some sort of an event system. Using random numbers to pull from a list of events to make each zone feel a bit more alive.