Combat Room Limiter and Cleanup
While adding monster movement, I ran into an issue. The user was able to spam the backwards reposition ability to do damage while avoiding all melee damage from the monster. To rectify this, I decided to add a room limiter.

This room limiter consists of a total of 21 spaces. The player will be loaded into a specific spot based on their preferred position, usually around relative positions 5-10. The monster spawns in the middle of the room. When a user uses a reposition ability, their relativePosition changes to reflect that.

When the user approaches the far left "wall" (relativePositions 0, and 1), the reposition abilities that move the character backwards will be removed from the combat ability list, so that the user may not continue moving backwards. This gives a bit more challenge, as once they hit the back wall, they can no longer spam the backwards reposition abilities to cheese the combat and are forced into moving forward and using chainers/finishers.

This implementation might have some issues later if I end up creating reposition abilities that move the user further than the clicked position, but at the time of commit, this is not the case and it works fine.

In addition to this implementation, I have added a visual representation of the room that the user and monster are in, so that the user may visualize where they are positioned in the room and know when they are getting close to the back wall.

console.logs have been cleaned up after this implementation
refactored the filler position objects into a variable that is now used throughout the calculateabilityposition function.