Celestial Tower Zone Inclusion, Refactoring of HUB, Fixed Longtime Worry
Celestial Tower is now fully implemented as a zone with world data, and is being passed the new redirect functions.

HUB has been cleaned up a ton and is now recieving most functions from Zone Controller.

I used to have an issue with being able to type in a path in a URL to "cheese" my way to a new location. With the new Zone Controller, that is no longer possible as it's all done via state.