Added New User Experience
I've added the NewUser component, which is only seen when the user logs in for the first time. It gives them an option to unlock all the basic abilities and sets their combat prefs for the weapon of their choice. This allows me to guaranteed that the user has combat prefs and other abilities unlocked prior to loading the first component, so that they may go directly to combat without any issues.

I will eventually flesh this out to add more weapons in the future, and possibly add the second weapon swap weapon later when that is implemented.