Fixed Combat Issues, Added Losing
Fixed having the user click on an empty skill in combat, resulting in an error. Having a skill position be empty means that you are not near that position and do not have a skill that can be used there.

Fixed health going below zero if hit for more damage than current health. Health now gets set to zero if the hit puts you under zero.

Added losing to combat. You can now lose!