Item Generation and Inventory Additions
Refactored item generation. Previously the item generation was a very ugly and lengthy if statement. It has been refactored into a loop to make it easier to read.

Added rarity bonus to item generation. This number will be an integer obtained by a calculation coming from the player and the monster/event killed or completed. If the player fails to roll the next rarity, it will remove the difference from the rarity bonus, and apply that rarity to the item. This happens until the bonus is used up.

Added amount variable to item generation. This allows the function to give multiple of one item to the character. This is only used for Tokens and Materials, so that I may give X amount of Cloth, Wood, Metal or Tokens to the player.

Added Tokens and Materials to item generation. They can now be obtained.

Temporary Monster Drop Tables added. This is currently implemented inside each monster's data. I'm not sure if this monster data will stay where it is, as it probably needs to be moved to the server.

Inventory can now properly view Tokens and Materials including their count number. The inventory panel needs an overhaul. Filters and Tabs are planned.